pyproject.toml
[tool.poetry]
name = "test-asigment"
version = "0.1.0"
description = ""
authors = ["Leonid3# <antonfenrisis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
argparse = "^1.4.0"
docker = "^7.0.0"
boto3 = "^1.34.72"
python-dotenv = "^1.0.1"
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
